Add missing evaluation filter (#8663)
* Add missing filter

* Update CHANGELOG

* Better solution reusing code

* Make vulnerabilityManagedFilters function common

* fix: wrong import

* fix: persist filters when switching tabs

* chore: remove unused imports

---------
