+++ plugins/main/public/components/overview/vulnerabilities/common/vulnerability-managed-filters.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import VulsEvaluationFilter, {
+  getUnderEvaluationFilterValue,
+  createUnderEvaluationFilter,
+  UNDER_EVALUATION_FIELD,
+} from './components/vuls-evaluation-filter';
+import { WAZUH_VULNERABILITIES_PATTERN } from '../../../../../common/constants';
+
+export const vulnerabilityManagedFilters = {
+  underEvaluation: {
+    managedField: UNDER_EVALUATION_FIELD,
+    selector: f =>
+      f.meta?.key === UNDER_EVALUATION_FIELD && f.meta?.type === 'phrase',
+    component: (props: {
+      managedFilter?: any;
+      setManagedFilter: (filters: any[]) => void;
+    }) => {
+      const onChange = (underEvaluation: boolean | null) => {
+        const newFilters = [];
+        if (underEvaluation !== null) {
+          newFilters.push(
+            createUnderEvaluationFilter(
+              underEvaluation,
+              WAZUH_VULNERABILITIES_PATTERN,
+            ),
+          );
+        }
+        props?.setManagedFilter(newFilters);
+      };
+
+      return (
+        <VulsEvaluationFilter
+          value={getUnderEvaluationFilterValue(props?.managedFilter)}
+          setValue={onChange}
+        />
+      );
+    },
+    order: 1,
+  },
+};
